pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vit-inductive-bias-distillation"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "torch==2.10.0",
    "torchvision==0.25.0",
    "torchmetrics==1.8.2",
    "numpy==2.4.2",
    "hydra-core>=1.3,<2.0",
    "datasets==4.6.1",
    "timm==1.0.24",
    "accelerate==1.12.0",
    "schedulefree==1.4.1",
]

[project.scripts]
basd-train = "src.train:main"
basd-eval = "src.eval:main"

[tool.setuptools.packages.find]
include = ["src*"]
